% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLFlow.R
\name{as_learner.OMLFlow}
\alias{as_learner.OMLFlow}
\title{Convert an OpenML Flow to a mlr3 Learner}
\usage{
\method{as_learner}{OMLFlow}(x, task_type = NULL, from_binary = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{task_type}{(\code{character(1)})
The task type to constrct a pseudo-learner. For more information see \link{OMLFlow}.}

\item{from_binary}{(\code{logical(1)}) Whether to construct the learner from the binary rds file
that is provided on OpenML. Note that when learners are constructed like this and the
versions of the installed packages don't match the flow's dependencies, the resulting
flow (and therefore also the resulting \link[mlr3:ResampleResult]{mlr3::ResampleResult}s) cannot be published on OpenML.
If it is FALSE, the function tries to construct the learner using \code{lrn()}.}
}
\description{
Convert an OpenML Flow to a mlr3 Learner
}
