% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{attach_to_study}
\alias{attach_to_study}
\title{Attach to a Study}
\usage{
attach_to_study(study_id, ids)
}
\arguments{
\item{study_id}{(\code{integer(1)})\cr
The id of the study.}

\item{ids}{(\code{integer()})\cr
The id of the object to attach to the study. Depending on the type of the study (it's main
entity type), this is either a task id or run id.}
}
\description{
Detaches an object from a study / collection.
}
