% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_design.R
\name{benchmark_design}
\alias{benchmark_design}
\title{Helper function to create a benchmark design}
\usage{
benchmark_design(tasks, learners, resamplings)
}
\arguments{
\item{tasks}{(\code{list()} or \code{Task}) A list of \link[mlr3:Task]{mlr3::Task}s.}

\item{learners}{(\code{list()} or \code{Learner}) A list of \link[mlr3:Learner]{mlr3::Learner}s.}

\item{resamplings}{(\code{list()} or \code{Resampling}) A list of \link[mlr3:Resampling]{mlr3::Resampling}s that are instantiated on the given tasks.}
}
\description{
\link{OMLTask}s contain tasks as well as resamplings. In order to create
a benchmark design from a list of tasks and corresponding instantiated resamplings, this
function can be used.
}
\examples{
\dontrun{
library("mlr3")
collection = OMLCollection$new(99)
otasks = collection$tasks[nrow <= 550, ][["task"]]
tasks = as_tasks(otasks)
resamplings = as_resamplings(otasks)
learners = lrns(c("classif.rpart", "classif.featureless"))
design = benchmark_design(tasks, learners, resamplings)
print(design)
bmr = benchmark(design)
}
}
