% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached}
\alias{cached}
\title{Cached}
\usage{
cached(fun, type, id, ..., cache_dir = FALSE)
}
\arguments{
\item{fun}{Download function, e.g. download_data_desc.}

\item{type}{The type that is downloaded, not really necessary}

\item{id}{The id of the object that is being downloaded, this can also be an url, as this is needed}
}
\description{
This function performs a cached version of the function 'fun'. I.e. it first checks whether
the objects is already stored in cache and returns it, otherwise it downloads it and stores it
in the cache folder.
}
