% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLFlow.R
\name{check_dependencies}
\alias{check_dependencies}
\title{When constructing the learner using lrn(), this function already indicates missing packages.
This function throws a warning in case the versions of the installed packages don't match the
dependencies of the flow on OpenML}
\usage{
check_dependencies(flow, verbose)
}
\description{
When constructing the learner using lrn(), this function already indicates missing packages.
This function throws a warning in case the versions of the installed packages don't match the
dependencies of the flow on OpenML
}
