% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{delete}
\alias{delete}
\title{Delete an object from OpenML}
\usage{
delete(type, id, api_key = NULL, server = NULL, confirm = TRUE)
}
\arguments{
\item{type}{(\code{character(1)})
What type of object should be deleted ("flow", "run", "task", "data", "study" / "collection")}

\item{id}{(\code{integer(1)})
The id of the object that will be deleted.}

\item{api_key}{(\code{character(1)})
The API key to perform the action, if left NULL it first tries getOption("mlr3oml.api_key") and
then Sys.getenv("OMLAPIKEY").}

\item{server}{(\code{character(1)})
The server address, defaults to \code{get_server()}.}

\item{confirm}{(\code{logical(1)})
Whether the deletion has to be confirmed interactively.}
}
\description{
This function can be used to delete objects from the OpenML server.
}
