% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{detach_from_study}
\alias{detach_from_study}
\title{Detach from a Study
Detaches an object from a study / collection.}
\usage{
detach_from_study(study_id, ids)
}
\arguments{
\item{study_id}{(\code{integer(1)})\cr
The id of the study.}

\item{ids}{(\code{integer()})\cr
The id of the object to attach to the study. Depending on the type of the study (it's main
entity type), this is either a task id or run id.}
}
\description{
Detach from a Study
Detaches an object from a study / collection.
}
