% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_autotest.R
\name{download_ids}
\alias{download_ids}
\title{make_download_test}
\usage{
download_ids(type, limit = 100000L)
}
\description{
Function factory that can create test functions for various OpenML objects.
}
\details{
This function creates a test object, that samples random ids and then tests
whether the downloaded OML objects satisfy the conditions defined in \code{expect_obj()}.
Because the sampled ids might not be existing OML objects, where is an additional
\code{download_error(id)} function, that takes as input the id and returns a \code{character(1)} that
contains the test message that is expected when the corresponding id does not exist on OpenML.
This error message has to be matched perfectly, otherwise the test fails.

The flow ids that are tested are sampled randomly. There are then two cases;:
\enumerate{
\item A flow with the id exists --> Has to satisfy expect_flow
\item A flow with the id does not exist --> We expect the correct 'download error'
This is done as long as either max_tries tries are exhausted or n flows that exist were found.
}
}
