% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_description.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Gets the dependencies in the form "mlr3_x.x.x, rpart_x.x.x" from the packages.}
\usage{
get_dependencies(x)
}
\description{
Gets the dependencies in the form "mlr3_x.x.x, rpart_x.x.x" from the packages.
}
