% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Gets the OpenML id of an mlr3 object (if it exists).}
\usage{
get_id(x, ...)
}
\arguments{
\item{x}{(\code{any}) The mlr3 object whose OpenML id is queried.}

\item{...}{Currently unused.}
}
\value{
Id of the object.
}
\description{
OpenML objects can be converted to mlr3 objects. After this conversion the id is
added to a private slot of the object. This S3 generic retries this id and checks whether the
object has been modified since conversion (by comparing the hashes). If it was modified
(hash changed), NULL is returned.
}
\examples{
\donttest{
otask = OMLTask$new(31)
task = as_task(otask)
# returns the id
get_id(task)
task$col_roles$features = "duration"
# returns NULL
get_id(task)
}

}
