% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_server}
\alias{get_server}
\title{Returns the currently active server}
\usage{
get_server()
}
\description{
Returns the server that can be configures via the option \code{mlr3oml.server}.
Returns "https://www.openml.org/api/v1" by default.
}
