% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_oml_data_sets.R
\name{list_oml_data_sets}
\alias{list_oml_data_sets}
\title{List Data Sets from OpenML}
\usage{
list_oml_data_sets(
  number_instances = NULL,
  number_features = NULL,
  number_classes = NULL,
  number_missing_values = NULL,
  tag = NULL,
  ...
)
}
\arguments{
\item{number_instances}{(\code{integer()})\cr
Filter for number of instances.}

\item{number_features}{(\code{integer()})\cr
Filter for number of features.}

\item{number_classes}{(\code{integer()})\cr
Filter for number of labels of the target (only classification tasks).}

\item{number_missing_values}{(\code{integer()})\cr
Filter for number of missing values.}

\item{tag}{(\code{character()})\cr
Filter for specific tag. You can provide multiple tags as character vector.}

\item{...}{(any)\cr
Additional filters as named arguments.}
}
\value{
(\code{data.table()}) of results.
}
\description{
This function allows to find data sets on \url{https://openml.org/d} using
some simple filter criteria.

Note that only a subset of filters are exposed here.
}
\details{
Filter values can be provided as single atomic values (typically integer or character).
Provide a numeric vector of length 2 (\code{c(l, u)}) to find matches in the range \eqn{[l, u]}.
}
\examples{
\dontrun{
list_oml_data_sets(number_instances = 150, number_features = c(1, 10))
}
}
