% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_oml_tasks.R
\name{list_oml_tasks}
\alias{list_oml_tasks}
\title{List Taks from OpenML}
\usage{
list_oml_tasks(...)
}
\arguments{
\item{...}{(named \code{character()})\cr
Named values of filter criteria, e.g. \code{tag = "study_99"}.}
}
\value{
(\code{data.table()}) of results.
}
\description{
This function is not yet finished.
You can pass filter criteria to construct an API request.
}
