% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_description.R
\name{make_description}
\alias{make_description}
\title{Create OpenmL xml description}
\usage{
make_description(x, ...)
}
\arguments{
\item{x}{(any) Object for which to make the xml description.}

\item{...}{Additional arguments.}
}
\description{
Creates the xml description for an mlr3 object so that it can be uploaded to
OpenML.
}
