% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_expectation.R
\name{make_expect}
\alias{make_expect}
\title{make_expect}
\usage{
make_expect(f, catch_id = TRUE)
}
\description{
Make a custom expectation function.
}
\details{
This function takes a function that consists of expectations in each line,
wraps them around show_failure and does one expect that checks that all tests
passed. If it does not it prints the failure messages that were collected along
the way. Because testthat and ceckmate expectations are mixed, we cannot rely on the
test_xxx function from checkmate but have to use expect + show_failure
}
