% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerOML.R
\name{make_oml_learner}
\alias{make_oml_learner}
\title{Creates a pseudo OpenML Learner
This object is used to create A learner for a flow that is not from R, but the results are used
in a resample or benchmark result}
\usage{
make_oml_learner(flow, task_type)
}
\description{
Creates a pseudo OpenML Learner
This object is used to create A learner for a flow that is not from R, but the results are used
in a resample or benchmark result
}
