% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_oml_prediction}
\alias{make_oml_prediction}
\title{When uploading predictions to OpenML, they have to satisfy a specific format.
This function takes a mlr3 Prediction and converts it into that format
TODO: check whether the -1L is really necessary for the fold_ids when the test server
is up and running again
Every OpenML prediction is a data.table with:
\link{repeat | fold | row_id | prediction | truth}
For classification tasks there are additional columns
\link{confidence.class1 | confidence.class2 | ...}
In mlr3 these are set to the probabilities if they are available and otherwise to 1 for the
predicted class and 0 for everything else.}
\usage{
make_oml_prediction(rr)
}
\description{
When uploading predictions to OpenML, they have to satisfy a specific format.
This function takes a mlr3 Prediction and converts it into that format
TODO: check whether the -1L is really necessary for the fold_ids when the test server
is up and running again
Every OpenML prediction is a data.table with:
\link{repeat | fold | row_id | prediction | truth}
For classification tasks there are additional columns
\link{confidence.class1 | confidence.class2 | ...}
In mlr3 these are set to the probabilities if they are available and otherwise to 1 for the
predicted class and 0 for everything else.
}
