% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_flow.R
\name{make_test_flows}
\alias{make_test_flows}
\title{Tests for OML Flows}
\usage{
make_test_flows()
}
\description{
Does various tests regarding upload, download and conversion of tests.
Some tests require the server, while others are simulated locally.
}
\details{
The following can tested:
\itemize{
\item Can flows be uploaded to the test server.
\item Can flows be downloaded from the server.
\item Can flows be converted to mlr3 learners.
For that locally stored files are used.
}

This writes test flows to ./inst/testdata that can be used threshold for local testing
}
