% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3oml-package}
\alias{mlr3oml}
\alias{mlr3oml-package}
\title{mlr3oml: Connector between 'mlr3' and 'OpenML'}
\description{
Provides an interface to 'OpenML.org' to download data and machine
  learning tasks. Tasks can be automatically converted to 'mlr3' tasks.
  For a more sophisticated interface which also supports listing and filtering
  of tasks as well as uploading experiments, see the 'OpenML' package.
}
\section{Options}{

\itemize{
\item \code{mlr3oml.cache}: Enables or disables caching globally.
If set to \code{FALSE}, caching is disabled.
If set to \code{TRUE}, cache directory as reported by \code{\link[=R_user_dir]{R_user_dir()}} is used.
Alternatively, you can specify a path on the local file system here.
Default is \code{FALSE}.
\item \code{mlr3oml.api_key}: API key to use.
}
}

\section{Logging}{

The \CRANpkg{lgr} package is used for logging.
To change the threshold, use \code{lgr::get_logger("mlr3oml")$set_threshold()}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3oml}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3oml}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})

}
