% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{oml_sugar}
\alias{oml_sugar}
\alias{oml_data}
\alias{oml_task}
\alias{oml_flow}
\alias{oml_run}
\alias{oml_collection}
\title{Syntactic Sugar for Object Construction}
\usage{
oml_data(
  id,
  cache = getOption("mlr3oml.cache", FALSE),
  parquet = getOption("mlr3oml.parquet", TRUE),
  test_server = getOption("mlr3oml.test_server", FALSE)
)

oml_task(
  id,
  cache = getOption("mlr3oml.cache", FALSE),
  parquet = getOption("mlr3oml.parquet", TRUE),
  test_server = getOption("mlr3oml.test_server", FALSE)
)

oml_flow(
  id,
  cache = getOption("mlr3oml.cache", FALSE),
  test_server = getOption("mlr3oml.test_server", FALSE)
)

oml_run(
  id,
  cache = getOption("mlr3oml.cache", FALSE),
  parquet = getOption("mlr3oml.parquet", TRUE),
  test_server = getOption("mlr3oml.test_server", FALSE)
)

oml_collection(
  id,
  cache = getOption("mlr3oml.cache", FALSE),
  parquet = getOption("mlr3oml.parquet", TRUE),
  test_server = getOption("mlr3oml.test_server", FALSE)
)
}
\arguments{
\item{id}{(\code{integer(1)})\cr
OpenML id for the object.}

\item{cache}{(\code{logical(1)} | \code{character(1)})\cr
See field \code{cache} for an explanation of possible values.
Defaults to value of option \code{"mlr3oml.cache"}, or \code{FALSE} if not set.}

\item{parquet}{(\code{logical(1)})\cr
Whether to use parquet instead of arff.
If parquet is not available, it will fall back to arff.}

\item{test_server}{(\code{character(1)})\cr
Whether to use the OpenML test server (https://test.openml.org/) or public server
(https://www.openml.org/).
Defaults to value of option \code{"mlr3oml.test_server"}, or \code{FALSE} if not set.}
}
\description{
Functions to create OpenML objects. The following functions are available:
\itemize{
\item \code{oml_data()} - creates an instance of the R6 class \link{OMLData}.
\item \code{oml_task()} - creates an instance of the R6 class \link{OMLTask}.
\item \code{oml_flow()} - creates an instance of the R6 class \link{OMLFlow}.
\item \code{oml_run()} - creates an instance of the R6 class \link{OMLRun}.
\item \code{oml_collection()} - creates an instance of the R6 class \link{OMLCollection}.
}
}
