% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLData.R
\name{OMLData}
\alias{OMLData}
\title{Interface to OpenML Data}
\description{
This is the class for data sets served on \url{https://openml.org/d}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{(\code{integer(1)})\cr
Data id on OpenML.}

\item{\code{use_cache}}{(\code{logical(1)})\cr
If \code{TRUE}, locally caches downloaded objects on the file system.file.
See \code{\link[=R_user_dir]{R_user_dir()}} for the location.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OMLData$new()}}
\item \href{#method-clone}{\code{OMLData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLData$new(id, use_cache = getOption("mlr3oml.use_cache", FALSE))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
