% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLContainer.R
\name{OMLDictionaryTask}
\alias{OMLDictionaryTask}
\title{OpenML Container}
\description{
Contains multiple OML objects and allows to convert them into a list of \link{mlr3} objects.
}
\references{
Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014).
\dQuote{OpenML.}
\emph{ACM SIGKDD Explorations Newsletter}, \bold{15}(2), 49--60.
\doi{10.1145/2641190.2641198}.
}
\section{Super class}{
\code{\link[mlr3oml:OMLDictionaryTask]{mlr3oml::OMLDictionaryTask}} -> \code{OMLDictionaryTask}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{OMLDictionaryTask$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3oml" data-topic="OMLDictionaryTask" data-id="add">}\href{../../mlr3oml/html/OMLDictionaryTask.html#method-add}{\code{mlr3oml::OMLDictionaryTask$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3oml" data-topic="OMLDictionaryTask" data-id="get">}\href{../../mlr3oml/html/OMLDictionaryTask.html#method-get}{\code{mlr3oml::OMLDictionaryTask$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3oml" data-topic="OMLDictionaryTask" data-id="initialize">}\href{../../mlr3oml/html/OMLDictionaryTask.html#method-initialize}{\code{mlr3oml::OMLDictionaryTask$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3oml" data-topic="OMLDictionaryTask" data-id="keys">}\href{../../mlr3oml/html/OMLDictionaryTask.html#method-keys}{\code{mlr3oml::OMLDictionaryTask$keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3oml" data-topic="OMLDictionaryTask" data-id="mget">}\href{../../mlr3oml/html/OMLDictionaryTask.html#method-mget}{\code{mlr3oml::OMLDictionaryTask$mget()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLDictionaryTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
