% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_flow.R
\name{pseudo_get_flow}
\alias{pseudo_get_flow}
\title{Loads the flows that were created with \code{pseudo_publish_learner} from disk into R.}
\usage{
pseudo_get_flow(name)
}
\description{
Loads the flows that were created with \code{pseudo_publish_learner} from disk into R.
}
