% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_flow.R
\name{pseudo_publish_learner}
\alias{pseudo_publish_learner}
\title{Creates the xml descriptin and writes it to the disk for testing purposes.}
\usage{
pseudo_publish_learner(x)
}
\description{
Creates the xml descriptin and writes it to the disk for testing purposes.
}
