% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publish on OpenML}
\usage{
publish(x, confirm = TRUE, ...)
}
\arguments{
\item{x}{(\code{any}) Object to publish to OpenML.}

\item{confirm}{(\code{logical(1)}) Whether to confirm the upload.}

\item{...}{Additional arguments.}
}
\description{
Publishes a  \link[mlr3:Learner]{mlr3::Learner} (as a Flow) or \link[mlr3:ResampleResult]{mlr3::ResampleResult} (as a Run) to OpenML.
Currently only resample results with atomic parameters can be uploaded, as it has to be
possible to represent the parameter values in json format.
}
\details{
This is a generic function.
}
