% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish.ResampleResult}
\alias{publish.ResampleResult}
\title{Publishes a ResampleResult}
\usage{
\method{publish}{ResampleResult}(x, upload_model = FALSE, ...)
}
\arguments{
\item{upload_model}{(\code{logical(1)}) Whether to upload the model with the resample result.}
}
\description{
Publishes a ResampleResult
}
