% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_collection.R
\name{publish_collection}
\alias{publish_collection}
\title{Publish a Collection to OpenML}
\usage{
publish_collection(
  ids,
  main_entity_type = "task",
  name,
  desc,
  alias = NULL,
  api_key = NULL,
  test_server = test_server_default()
)
}
\arguments{
\item{ids}{(\code{integer()})\cr
The IDs to include in the collection.}

\item{main_entity_type}{(\code{character(1)})\cr
The main entity type of the collection. Can be either "task" or "run".}

\item{name}{(\code{character(1)})\cr
The name for the collection.}

\item{desc}{(\code{character(1)})\cr
The description of the collection.}

\item{alias}{(\code{chraacter(1)})\cr
The alias for the collection.}

\item{test_server}{(\code{character(1)})\cr
Whether to use the OpenML test server or public server.
Defaults to value of option \code{"mlr3oml.test_server"}, or \code{FALSE} if not set.}
}
\description{
Publish a collection to OpenML
}
