% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLRun.R
\name{split_predictions}
\alias{split_predictions}
\title{splits the predictions from OpenML into mlr3 readable forat}
\usage{
split_predictions(predictions, resampling, task_type)
}
\description{
splits the predictions from OpenML into mlr3 readable forat
}
