% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_autotest.R
\name{unlist_leafs}
\alias{unlist_leafs}
\title{Autotest for flows}
\usage{
unlist_leafs(l, i = 1)
}
\description{
Does various tests regarding upload, download and conversion of tests.
}
\details{
The following can tested:
\itemize{
\item Can flows be uploaded to the test server.
\item Can flows be downloaded from the server.
\item Can flows be converted to mlr3 learners.
For that locally stored files are used.
}

Helper function when converting xml-document into a list
Turns list(a = list(1), b = list(c = list(2))) into list(a = 1, b = list(c = 2))
}
