% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_test_flow.R
\name{unlist_leaves}
\alias{unlist_leaves}
\title{Helper function when creating the list from the xml-description
When applied to list(a = list(1), b = list(c = list(2))) it returns
list(a = 1, b = list(2))}
\usage{
unlist_leaves(l, i = 1)
}
\description{
Helper function when creating the list from the xml-description
When applied to list(a = list(1), b = list(c = list(2))) it returns
list(a = 1, b = list(2))
}
