% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{url_to_filename}
\alias{url_to_filename}
\title{When an url is passed to the download function, this constructs the name how the element is
saved when caching is enabled}
\usage{
url_to_filename(url)
}
\description{
When an url is passed to the download function, this constructs the name how the element is
saved when caching is enabled
}
