% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_arff.R
\name{write_arff}
\alias{write_arff}
\title{Write ARFF files}
\usage{
write_arff(data, path, relation = deparse(substitute(data)))
}
\arguments{
\item{data}{(\code{data.frame()})\cr
Data to write.}

\item{path}{(\code{character(1)})\cr
Path or URI of the ARFF file, passed to \code{\link[=file]{file()}}.}

\item{relation}{(\code{character(1)})\cr
Relation (name) of the data set.}
}
\description{
Writes a \code{data.frame()} to an ARFF file.

Limitations:
\itemize{
\item Logicals are written as categorical features.
\item \link{POSIXct} columns are converted to UTC.
}
}
