% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLFlow.R
\name{as_learner.OMLFlow}
\alias{as_learner.OMLFlow}
\title{Convert an OpenML Flow to a mlr3 Learner}
\usage{
\method{as_learner}{OMLFlow}(x, task_type = NULL, from_binary = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{(OMLFlow) The OMLFlow that is converted to a mlr3::Learner.}

\item{task_type}{(\code{character(1)})
The task type to constrct a pseudo-learner. For more information see \link{OMLFlow}.}

\item{from_binary}{(\code{logical(1)}) Whether to construct an actual mlr3 learner.
flow (works of course only for mlr3 Flows).}

\item{verbose}{(\code{logical(1)}) Whether to be verbose.}

\item{...}{Additional arguments}
}
\description{
By default this function creates a Pseudo-Learner (that cannot be used for training or
prediction) for the given task type. This enables the conversion of OpenML Runs to
\link[mlr3:ResampleResult]{mlr3::ResampleResult}s.
The parameters of the flow include the parameters of all subcomponents, where the name
c.123.par means that this is the parameter 'par' of the component with id 123.
This is well defined because each ids can only appear once in a Flow.
}
