% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLFlow.R
\name{as_learner.OMLFlow}
\alias{as_learner.OMLFlow}
\title{Convert an OpenML Flow to a mlr3 Learner}
\usage{
\method{as_learner}{OMLFlow}(x, task_type = NULL, verbose = TRUE, ...)
}
\arguments{
\item{task_type}{(\code{character(1)})
The task type to constrct a pseudo-learner. For more information see \link{OMLFlow}.}

\item{verbose}{(\code{logical(1)}) Whether to give informative print output}
}
\description{
If it is not an mlr3 Flow it creates a pseudo learner (if \code{task_type} is given, otherwise
it returns NULL).
If it is a mlr3 Flow it downloads the binary rds file and compares the dependencies of the
flow with the installed packages and the running R version (gives informative print output if
the verbose flag is set).
}
