% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_design.R
\name{benchmark_design}
\alias{benchmark_design}
\title{Helper function to create a benchmark design}
\usage{
benchmark_design(tasks, learners, resamplings)
}
\arguments{
\item{tasks}{list of \link[mlr3:Task]{mlr3::Task}s}

\item{learners}{list of \link[mlr3:Learner]{mlr3::Learner}s}

\item{resamplings}{list of \link[mlr:Resampling]{mlr::Resampling}s for the tasks.}
}
\description{
\link{OMLTasks}s contain tasks as well as resamplings. In order to create
a benchmark design from a list of tasks and corresponding instantiated resamplings, this
function can be used.
}
