% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cached}
\alias{cached}
\title{Cached}
\usage{
cached(fun, type, id, ..., cache_dir = FALSE)
}
\arguments{
\item{fun}{Download function}

\item{type}{The type that is downloaded, not really necessary}
}
\description{
This function
}
