% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check the dependencies of the flow and compare it with the installed versions}
\usage{
check_dependencies(flow, verbose)
}
\arguments{
\item{flow}{(\code{OMLFLow}) The flow whose dependencies are checked. Should only  be called on
mlr3 Flows.}

\item{versbose}{(\code{logical(1)}) Whether to be verbose.}
}
\value{
Returns TRUE if the dependencies match and FALSE otherwise
}
\description{
Compares the dependencies of the flow with those installed.
}
