% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerOML.R
\name{construct_paramset}
\alias{construct_paramset}
\title{this constructs a paramset for the parameter field of a OMLFlow}
\usage{
construct_paramset(parameter)
}
\description{
this constructs a paramset for the parameter field of a OMLFlow
}
