% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLDictionary.R
\name{OMLDictionary}
\alias{OMLDictionary}
\title{OpenML Dictionary}
\description{
OpenML Dictionary

OpenML Dictionary
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OMLDictionary$new()}}
\item \href{#method-get}{\code{OMLDictionary$get()}}
\item \href{#method-keys}{\code{OMLDictionary$keys()}}
\item \href{#method-mget}{\code{OMLDictionary$mget()}}
\item \href{#method-add}{\code{OMLDictionary$add()}}
\item \href{#method-clone}{\code{OMLDictionary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLDictionary$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
get an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLDictionary$get(key, convert = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Can either be the name (data) of the task or the id.}

\item{\code{convert}}{Whether the resulting object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keys"></a>}}
\if{latex}{\out{\hypertarget{method-keys}{}}}
\subsection{Method \code{keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLDictionary$keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mget"></a>}}
\if{latex}{\out{\hypertarget{method-mget}{}}}
\subsection{Method \code{mget()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLDictionary$mget(keys = NULL, convert = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLDictionary$add(task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLDictionary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
