% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLFlow.R
\name{OMLFlow}
\alias{OMLFlow}
\title{Interface to OpenML Flows}
\description{
This is the class for flows provided on the \href{https://new.openml.org/search?type=flow&sort=runs}{OpenML website}.
}
\examples{
\donttest{
flow = OMLFlow$new(id = 1)
}
}
\references{
Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014).
\dQuote{OpenML.}
\emph{ACM SIGKDD Explorations Newsletter}, \bold{15}(2), 49--60.
\doi{10.1145/2641190.2641198}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{(\code{integer(1)})\cr
OpenML flow id.}

\item{\code{cache_dir}}{(\code{logical(1)} | \code{character(1)})\cr
Stores the location of the cache for objects retrieved from \url{https://openml.org}.
If set to \code{FALSE}, caching is disabled.

The package \CRANpkg{qs} is required for caching.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{desc}}{(\code{list(n)})\cr
The description as downloaded from OpenML.}

\item{\code{parameter}}{(\code{data.table})\cr
The parameters of the flow.}

\item{\code{tag}}{(\code{character(n)})\cr
The tags of the flow.}

\item{\code{dependencies}}{(\code{character(n)})\cr
The dependencies of the flow.}

\item{\code{name}}{(\code{character(1)})\cr
The name of the flow.}

\item{\code{description}}{(\code{character(1)})\cr
The description of the flow.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OMLFlow$new()}}
\item \href{#method-convert}{\code{OMLFlow$convert()}}
\item \href{#method-print}{\code{OMLFlow$print()}}
\item \href{#method-clone}{\code{OMLFlow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new object of class \link{OMLFlow}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLFlow$new(id, cache = getOption("mlr3oml.cache", FALSE))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert"></a>}}
\if{latex}{\out{\hypertarget{method-convert}{}}}
\subsection{Method \code{convert()}}{
Tries to convert the OMLFlow into an \link[mlr3:Learner]{mlr3::Learner} if the name starts with mlr3.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLFlow$convert()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLFlow$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLFlow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
