% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLResampling.R
\name{OMLResampling}
\alias{OMLResampling}
\title{Interface to OpenML Splits}
\description{
This is the class for OpenML splits that belong to task.
}
\section{mlr3Intergration}{

Intermediary object that can be converted to mlr3::Resampling.
}

\examples{
\donttest{
otask = OMLTask$new(59) # mlr3oml::OMLTask
oresampling = otask$resampling # [mlr3oml::OMLResampling]
resampling = oresampling$convert() # mlr3::Resampling
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OMLResampling$new()}}
\item \href{#method-convert}{\code{OMLResampling$convert()}}
\item \href{#method-clone}{\code{OMLResampling$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLResampling$new(
  task = NULL,
  task_id = NULL,
  cache = getOption("mlr3oml.cache", FALSE)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert"></a>}}
\if{latex}{\out{\hypertarget{method-convert}{}}}
\subsection{Method \code{convert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLResampling$convert()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLResampling$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
