% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLTask.R
\name{OMLTask}
\alias{OMLTask}
\title{Interface to OpenML Tasks}
\description{
This is the class for data sets served on \url{https://openml.org/t}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{(\code{integer(1)})\cr
OpenML task id.}

\item{\code{use_cache}}{(\code{logical(1)})\cr
If \code{TRUE}, locally caches downloaded objects on the file system.file.
See \code{\link[=R_user_dir]{R_user_dir()}} for the path (depending on your operating system).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OMLTask$new()}}
\item \href{#method-clone}{\code{OMLTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new object of class \code{OMLData}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLTask$new(id, use_cache = getOption("mlr3oml.use_cache", FALSE))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{integer(1)})\cr
OpenML task id.}

\item{\code{use_cache}}{(\code{logical(1)})\cr
Flag to control the file system cache.
See \code{\link[=R_user_dir]{R_user_dir()}} for the path (depending on your operating system).
Default is the value of option \code{"mlr3oml.use_cache"} or \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
