% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMLTask.R
\name{OMLTask}
\alias{OMLTask}
\title{Interface to OpenML Tasks}
\description{
This is the class for tasks served on \href{https://openml.org/search?type=task&sort=runs}{OpenML}.
It consists of a dataset and other meta-information such as the target variable for supervised
problems.
}
\section{mlr3 Integration}{

\itemize{
\item Obtain a \link[mlr3:Task]{mlr3::Task} by calling \code{as_task()}.
\item Obtain a \link[mlr3:Resampling]{mlr3::Resampling} by calling \code{as_resampling()}.
}
}

\examples{
library("mlr3")
\donttest{
# Get a task from OpenML:
otask = OMLTask$new(id = 31L)
otask$data
otask$data_split
otask$target_names
otask$feature_names

# convert to mlr3 Task:
task = as_task(otask)

# get a task via tsk():
tsk("oml", task_id = 31L)
}
}
\references{
Vanschoren J, van Rijn JN, Bischl B, Torgo L (2014).
\dQuote{OpenML.}
\emph{ACM SIGKDD Explorations Newsletter}, \bold{15}(2), 49--60.
\doi{10.1145/2641190.2641198}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{(\code{integer(1)})\cr
OpenML task id.}

\item{\code{cache_dir}}{(\code{logical(1)} | \code{character(1)})\cr
Stores the location of the cache for objects retrieved from \url{https://openml.org}.
If set to \code{FALSE}, caching is disabled.

The package \CRANpkg{qs} is required for caching.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{(\code{character(1)})\cr
Name of the task, extracted from the task description.}

\item{\code{task_type}}{(\code{character(1)})\cr
The OpenML task type.}

\item{\code{desc}}{(\code{list()})\cr
Task description (meta information), downloaded and converted from the JSON API response.}

\item{\code{data_id}}{(\code{integer()})\cr
Data id, extracted from the task description.}

\item{\code{data}}{(\link{OMLData})\cr
Access to the underlying OpenML data set via a \link{OMLData} object.}

\item{\code{nrow}}{(\code{integer()})\cr
Number of rows, extracted from the \link{OMLData} object.}

\item{\code{ncol}}{(\code{integer()})\cr
Number of columns, as extracted from the \link{OMLData} object.}

\item{\code{target_names}}{(\code{character()})\cr
Name of the targets, as extracted from the OpenML task description.}

\item{\code{feature_names}}{(\code{character()})\cr
Name of the features (without targets of this \link{OMLTask}).}

\item{\code{data_split}}{(\link{OMLDataSplit})\cr
Access to the OpenML Evaluation Procedure.}

\item{\code{tags}}{(\code{character()})\cr
Returns all tags of the task.}

\item{\code{data_name}}{(\code{character()})\cr
Name of the dataset (inferred from the task name).}

\item{\code{parquet}}{(\code{logical(1)})\cr
Whether to use parquet.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMLTask-new}{\code{OMLTask$new()}}
\item \href{#method-OMLTask-print}{\code{OMLTask$print()}}
\item \href{#method-OMLTask-clone}{\code{OMLTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMLTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-OMLTask-new}{}}}
\subsection{Method \code{new()}}{
Creates a new object of class \code{OMLTask}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLTask$new(
  id,
  cache = getOption("mlr3oml.cache", FALSE),
  parquet = getOption("mlr3oml.parquet", FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{integer(1)})\cr
OpenML task id.}

\item{\code{cache}}{(\code{logical(1)} | \code{character(1)})\cr
See field \code{cache} for an explanation of possible values.
Defaults to value of option \code{"mlr3oml.cache"}, or \code{FALSE} if not set.}

\item{\code{parquet}}{(\code{logical(1)})\cr
Whether to use parquet instead of arff.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMLTask-print"></a>}}
\if{latex}{\out{\hypertarget{method-OMLTask-print}{}}}
\subsection{Method \code{print()}}{
Prints the object.
For a more detailed printer, convert to a \link[mlr3:Task]{mlr3::Task} via \verb{$task}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLTask$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMLTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMLTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMLTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
