% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publish on OpenML}
\usage{
publish(x, confirm = TRUE, ...)
}
\arguments{
\item{x}{(\code{any}) Object to publish to OpenML.}

\item{confirm}{(\code{logical(1)}) Whether to confirm the upload.}

\item{...}{Additional arguments.}
}
\description{
Publishes a {Flow, Run, Dataset, Task} to OpenML.
}
\details{
This is a generic function.
}
