% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_collection.R
\name{publish_collection}
\alias{publish_collection}
\title{Publish a Collection to OpenML}
\usage{
publish_collection(
  ids,
  main_entity_type = "task",
  name,
  desc,
  alias = NULL,
  api_key = NULL,
  test_server = test_server_default(),
  ...
)
}
\arguments{
\item{ids}{(\code{integer()})\cr
The IDs to include in the collection.}

\item{main_entity_type}{(\code{character(1)})\cr
The main entity type of the collection. Can be either "task" or "run".}

\item{name}{(\code{character(1)})\cr
The name of the collection.}

\item{desc}{(\code{character(1)})\cr
The descrition of the collection.}

\item{alias}{(\code{chraacter(1)})\cr
The alias of the collection.}
}
\description{
Publish a collectio to OpenML
}
